move sb_connect  to utils/database
